+++ src/ReleaseNotes.md
@@ -5,6 +5,8 @@ Bugfix release increasing stability.
 - check if cURL is installed when installing .NET or vsdbg
 - fix analyzer warnings
 - added message how to add .NET to search path
+- added support to publish app as self contained apps
+- added support for debugging self contained apps
 
 ## Bug Fixes
 - fixed caption for error messages